refactor: divide app by modules
