+++ src/app/modules/header/header.component.scss
@@ -1,4 +1,4 @@
-@use '~@angular/material' as mat;
+@use '../../../../node_modules/@angular/material/index' as mat;
 
 .header {
   min-height: 80px;
