+++ src/app/modules/header/search-bar/search-bar.component.scss
