+++ src/app/modules/header/sorting-settings/filter-by-word/filter-by-word.component.scss
@@ -1,4 +1,4 @@
-@use '~@angular/material' as mat;
+@use '../../../../../../node_modules/@angular/material/index' as mat;
 
 .word-input {
   height: 30px;
