+++ src/app/modules/header/sorting-settings/sorting-settings.component.scss
