+++ src/app/modules/header/user-profile/user-profile.component.scss
