+++ src/app/modules/results-list/directives/date-border.directive.ts
