+++ src/app/modules/results-list/results-item/results-item.component.scss
