+++ src/app/modules/results-list/results-list.component.scss
