+++ src/app/shared/pipes/trim-text.pipe.ts
